# Sparsity-preserving outer products