--- a/stdlib/SparseArrays/src/higherorderfns.jl
+++ b/stdlib/SparseArrays/src/higherorderfns.jl
@@ -93,9 +93,8 @@ is_supported_sparse_broadcast(t::Union{Transpose, Adjoint}, rest...) = is_suppor
 is_supported_sparse_broadcast(x, rest...) = axes(x) === () && is_supported_sparse_broadcast(rest...)
 is_supported_sparse_broadcast(x::Ref, rest...) = is_supported_sparse_broadcast(rest...)
 
-is_specialcase_sparse_broadcast(f, rest...) = false
-is_specialcase_sparse_broadcast(::typeof(*), ::SparseVectorUnion,
-                                ::AdjOrTransSparseVectorUnion) = true
+can_skip_sparsification(f, rest...) = false
+can_skip_sparsification(::typeof(*), ::SparseVectorUnion, ::AdjOrTransSparseVectorUnion) = true
 
 # Dispatch on broadcast operations by number of arguments
 const Broadcasted0{Style<:Union{Nothing,BroadcastStyle},Axes,F} =
@@ -817,33 +816,32 @@ _finishempty!(C::SparseMatrixCSC) = (fill!(C.colptr, 1); C)
 
 # special case - vector outer product
 _copy(f::typeof(*), x::SparseVectorUnion, y::AdjOrTransSparseVectorUnion) = _outer(x, y)
-@inline _outer(x::SparseVectorUnion, y::Adjoint) = return _outer(conj, x, y)
-@inline _outer(x::SparseVectorUnion, y::Transpose) = return _outer(identity, x, y)
+@inline _outer(x::SparseVectorUnion, y::Adjoint) = return _outer(conj, x, parent(y))
+@inline _outer(x::SparseVectorUnion, y::Transpose) = return _outer(identity, x, parent(y))
 function _outer(trans::Tf, x, y) where Tf
-    w = parent(y)
     nx = length(x)
-    nw = length(w)
+    ny = length(y)
     rowvalx = nonzeroinds(x)
-    rowvalw = nonzeroinds(w)
+    rowvaly = nonzeroinds(y)
     nzvalsx = nonzeros(x)
-    nzvalsw = nonzeros(w)
+    nzvalsy = nonzeros(y)
     nnzx = length(nzvalsx)
-    nnzw = length(nzvalsw)
+    nnzy = length(nzvalsy)
 
-    nnzC = nnzx * nnzw
-    Tv = typeof(oneunit(eltype(x)) * oneunit(eltype(w)))
-    Ti = promote_type(indtype(x), indtype(w))
-    colptrC = zeros(Ti, nw + 1)
+    nnzC = nnzx * nnzy
+    Tv = typeof(oneunit(eltype(x)) * oneunit(eltype(y)))
+    Ti = promote_type(indtype(x), indtype(y))
+    colptrC = zeros(Ti, ny + 1)
     rowvalC = Vector{Ti}(undef, nnzC)
     nzvalsC = Vector{Tv}(undef, nnzC)
 
     idx = 0
     @inbounds colptrC[1] = 1
-    @inbounds for jj = 1:nnzw
-        wval = nzvalsw[jj]
-        iszero(wval) && continue
-        col = rowvalw[jj]
-        wval = trans(wval)
+    @inbounds for jj = 1:nnzy
+        yval = nzvalsy[jj]
+        iszero(yval) && continue
+        col = rowvaly[jj]
+        yval = trans(yval)
 
         for ii = 1:nnzx
             xval = nzvalsx[ii]
@@ -851,12 +849,12 @@ function _outer(trans::Tf, x, y) where Tf
             idx += 1
             colptrC[col+1] += 1
             rowvalC[idx] = rowvalx[ii]
-            nzvalsC[idx] = xval * wval
+            nzvalsC[idx] = xval * yval
         end
     end
     cumsum!(colptrC, colptrC)
 
-    return SparseMatrixCSC(nx, nw, colptrC, rowvalC, nzvalsC)
+    return SparseMatrixCSC(nx, ny, colptrC, rowvalC, nzvalsC)
 end
 
 # (9) _broadcast_zeropres!/_broadcast_notzeropres! for more than two (input) sparse vectors/matrices
@@ -1127,7 +1125,7 @@ broadcast(f::Tf, A::SparseMatrixCSC, ::Type{T}) where {Tf,T} = broadcast(x -> f(
 
 function copy(bc::Broadcasted{PromoteToSparse})
     bcf = flatten(bc)
-    if is_specialcase_sparse_broadcast(bcf.f, bcf.args...)
+    if can_skip_sparsification(bcf.f, bcf.args...)
         return _copy(bcf.f, bcf.args...)
     elseif is_supported_sparse_broadcast(bcf.args...)
         return broadcast(bcf.f, map(_sparsifystructured, bcf.args)...)