@jmert 2017-12-08
Looks like the first commit created a method ambiguity, and I might want to work on an implementation which will rebase on top of #24969 easily.