@jmert 2018-12-18
As a reference point for myself — current master (v1.2.0-DEV.27) using similar benchmarks as the original post destroys my computer for a few minutes. I think the second `@benchmark`'s timing is inflated because I actually run out of RAM and start hitting swap (i.e. during the second benchmark, `htop` output showed all 16GB of RAM on my laptop being used, and ~57 GB of virtual memory mapped, with periodic stalls as my computer becomes overtaxed.)

```
  1.998 s (46 allocations: 37.25 GiB)
  43.104 s (4 allocations: 18.63 GiB)
  945.432 ms (26 allocations: 904.92 MiB)
  6.910 s (6 allocations: 4.66 GiB)
```