@jmert 2018-12-26
I've reimplemented this PR on the current master. The big differences from before are:
1. The core optimization is implemented within the broadcasting code path rather than as specific methods of `kron` and `*`. Unfortunately, I don't really understand the sparse broadcasting all that well, so adding in this special case is done through a sort of hack. (See `is_specialcase_sparse_broadcast()`.)
2. No specialization occurs for mixed dense-sparse outer products (yet).

The updated benchmark I'm working with is:
```julia
using BenchmarkTools, SparseArrays, Test

m,n = (2_000, 10_000);
k = 1000;
A = sprand(m, n, 0.01);
a = view(A, :, k);
u = A[:, k];
x = sparse(rand(n));
X = sparse(collect(1:n), fill(k, n), x, n, n);

suite = BenchmarkGroup()
# outer products only
suite["x × u'"] = @benchmarkable $x * $(u');
suite["x × a'"] = @benchmarkable $x * $(a');
# quadratic matrix products which motivated
suite["A × X × A'"] = @benchmarkable $A * $X * $(A');
suite["(A × x) × a'"] = @benchmarkable ($A * $x) * $(a');
tune!(suite)

before = run(suite)

using Revise
Revise.track(Base)

after = run(suite)

if true
    println("Before vs after for specific operations")
    println("=======================================\n")
    for key in keys(after)
        t1, t0 = minimum.((after[key], before[key]))
        r = judge(t1, t0)
        print(key, ": ")
        show(stdout, MIME"text/plain"(), r)
        println()
    end

    println("\n\n")
    println("Optimized quadratic product, before and after")
    println("=============================================\n")
    for (trial,name) in [(before,"before"), (after,"after")]
        key1 = "(A × x) × a'"
        key2 = "A × X × A'"
        t1, t0 = minimum.((trial[key1], trial[key2]))
        r = judge(t1, t0)
        print("$key1 vs $key2 ($name): ")
        show(stdout, MIME"text/plain"(), r)
        println()
    end
end
```

and gives results

```julia-repl
Before vs after for specific operations
=======================================

(A × x) × a': BenchmarkTools.TrialJudgement:
  time:   -99.82% => improvement (5.00% tolerance)
  memory: -97.70% => improvement (1.00% tolerance)
x × a': BenchmarkTools.TrialJudgement:
  time:   -99.96% => improvement (5.00% tolerance)
  memory: -97.89% => improvement (1.00% tolerance)
A × X × A': BenchmarkTools.TrialJudgement:
  time:   +0.38% => invariant (5.00% tolerance)
  memory: +0.00% => invariant (1.00% tolerance)
x × u': BenchmarkTools.TrialJudgement:
  time:   -98.03% => improvement (5.00% tolerance)
  memory: -98.95% => improvement (1.00% tolerance)



Optimized quadratic product, before and after
=============================================

(A × x) × a' vs A × X × A' (before): BenchmarkTools.TrialJudgement:
  time:   +6941.20% => regression (5.00% tolerance)
  memory: +290.58% => regression (1.00% tolerance)
(A × x) × a' vs A × X × A' (after): BenchmarkTools.TrialJudgement:
  time:   -87.66% => improvement (5.00% tolerance)
  memory: -91.02% => improvement (1.00% tolerance)
```