@ViralBShah 2019-01-03
The outer product part is straightforward. If @mbauman or @KristofferC can check the broadcasting bits, we can probably merge. @andreasnoack mentioned about `lowrankupdate!` and I am not sure if we want to do anything further.

Also, it is fine to get this performance improvement in and do further improvements in other PRs.