@mbauman 2019-01-03
Am I correct in understanding that the only functional change here is that `kron` used to return a dense matrix whereas now it's appropriately sparse?  Everything else is performance.