@jmert 2019-01-04
> @andreasnoack mentioned about lowrankupdate! and I am not sure if we want to do anything further.

Correct. The old PR had put in a stub function, but any performance "improvement" implied by providing an inplace update would have been a lie since I wasn't implementing such a feature at that time. If someone wants to tackle that, I agree that it can be in a new PR.

> I'm impressed — the sparse broadcasting machinery has survived a quite a few incremental code changes and is a bit of a tangled mess. You've adeptly hooked in right where it'd make sense.

Thanks!! I did have to abandon several approaches after realizing it wasn't going to work out as I'd hoped.

> Am I correct in understanding that the only functional change here is that kron used to return a dense matrix whereas now it's appropriately sparse? Ah, and also multiplication with views. Everything else is performance.

@ViralBShah has the current story correct — since I started this PR a year ago, the switch to `Transpose`/`Adjoint` has caused a huge slow-down by calculating the outer product via the dense `LinearAlgebra` generic code. This PR now maintains sparsity in `kron(u, v)`, `u * v'`, and `u .* v'`.

After the rewrite to current master, I dropped one of my original goals — having dense-sparse outer products produce a sparse output. I decided I was unsure enough of the broadcasting implementation at this point to spend time working on that goal; issue #26613 would be a good place to revive that conversation and let those enhancements occur in synchrony with a future PR to extend this implementation.