@jmert 2019-01-05
> Can you rebase on master (unless you did it in the last day or two) so that we pick up all the CI fixes?

No problem — done.