@KristofferC 2019-01-10
> Can you rebase on master (unless you did it in the last day or two) so that we pick up all the CI fixes?

Note that CI runs on the merge commit so rebasing doesn't really do anything w.r.t to CI.

I don't really see how this can be backported to a patch release since it is changing behavior? Tentatively removing backport label. Also, this needs a NEWS entry.