@abraunst 2019-01-19
Should `kron(::Diagonal, x::Union{SparseVector, SparseMatrix})` return `SparseMatrixCSC`? (sure!)
Should `kron(::Diagonal, x::Matrix)` return `SparseMatrixCSC`? (this would be consistent with this PR -- but currently there is an implementation for diagonal times dense...)